refactor: /shared workflows folder
