+++ src/app/features/workflows/shared/fasta.utils.ts
